# apps/api/openapi/school.json
{
  "components": {
    "schemas": {
...
      "ApiErrorBodyDto": {
        "properties": {
          "code": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ]
          },
          "details": {
            "additionalProperties": true,
            "nullable": true,
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "requestId"
        ],
        "type": "object"
      },
      "ApiErrorDto": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBodyDto"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "ArchiveDto": {
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 3,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CreateSubjectDto": {
        "properties": {
          "code": {
            "description": "Trimmed and upper-cased",
            "nullable": true,
            "pattern": "^[A-Z0-9-]{1,20}$",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "ErrorCode": {
        "enum": [
          "MALFORMED_REQUEST",
          "AUTH_REQUIRED",
          "AUTH_FAILED",
          "PERMISSION_DENIED",
          "SCHOOL_SUSPENDED",
          "ORIGIN_REJECTED",
          "NOT_FOUND",
          "VALIDATION_FAILED",
          "REFERENCE_NOT_FOUND",
          "UNKNOWN_FIELD",
          "INVALID_VALUE",
          "PAYLOAD_TOO_LARGE",
          "UNSUPPORTED_MEDIA_TYPE",
          "RATE_LIMITED",
          "SERVICE_UNAVAILABLE",
          "INTERNAL_ERROR",
          "PASSWORD_CHANGE_REQUIRED",
          "TOTP_REQUIRED",
          "TOTP_NOT_ENROLLED",
          "TOTP_ALREADY_ENROLLED",
          "TOTP_INVALID",
          "CURRENT_PASSWORD_INCORRECT",
          "SCHOOL_SHORT_CODE_TAKEN",
          "SCHOOL_SHORT_CODE_IMMUTABLE",
          "SCHOOL_TERMINATED",
          "ILLEGAL_STATUS_TRANSITION",
          "CONCURRENT_UPDATE",
          "TOKEN_INVALID",
          "EMAIL_NOT_VERIFIED",
          "SELF_ACTION_FORBIDDEN",
          "LAST_PRINCIPAL",
          "IDENTITY_NUMBER_MISSING",
          "ACTIVE_PRINCIPAL_EXISTS",
          "ALREADY_PRINCIPAL",
          "STAFF_NOT_ACTIVE",
          "USER_DISABLED",
          "ACADEMIC_YEAR_NAME_TAKEN",
          "ACADEMIC_YEAR_CLOSED",
          "ACADEMIC_YEAR_HAS_ACTIVE_ENROLMENTS",
          "CLASS_NAME_TAKEN",
          "CLASS_YEAR_IMMUTABLE",
          "CLASS_ARCHIVED",
          "CLASS_HAS_ACTIVE_ENROLMENTS",
          "SECTION_NAME_TAKEN",
          "SECTION_ARCHIVED",
          "SECTION_IN_USE",
          "SUBJECT_NAME_TAKEN",
          "SUBJECT_CODE_TAKEN",
          "SUBJECT_ARCHIVED",
          "GUARDIAN_CNIC_EXISTS",
          "GUARDIAN_CNIC_LOCKED",
          "GUARDIAN_MERGED",
          "GUARDIAN_CNIC_MISSING",
          "GUARDIAN_NO_LOGIN_LINK",
          "GUARDIAN_IS_PRIMARY_CONTACT",
          "LOGIN_ALREADY_EXISTS",
          "LINK_EXISTING_LOGIN_UNCONFIRMED",
          "UNEXPECTED_RESPONSE"
        ],
        "type": "string"
      },
...
      "SubjectDto": {
        "properties": {
          "archivedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "code": {
            "nullable": true,
            "pattern": "^[A-Z0-9-]{1,20}$",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "pattern": "^[1-9][0-9]{0,18}$",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "code",
          "archivedAt",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "SubjectSort": {
        "enum": [
          "name",
          "-name",
          "code",
          "-code"
        ],
        "type": "string"
      },
...
      "UpdateSubjectDto": {
        "properties": {
          "code": {
            "description": "Trimmed and upper-cased; null clears",
            "nullable": true,
            "pattern": "^[A-Z0-9-]{1,20}$",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "ASMS school API",
    "version": "1"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/subjects": {
      "get": {
        "operationId": "SubjectsController_list",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 999999,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 50,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Include archived subjects",
            "in": "query",
            "name": "includeArchived",
            "required": false,
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "description": "Name contains (case-insensitive), or code starts with",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 50,
              "minLength": 2,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SubjectSort"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/SubjectDto"
                      },
                      "type": "array"
                    },
                    "limit": {
                      "maximum": 50,
                      "minimum": 1,
                      "type": "integer"
                    },
                    "page": {
                      "minimum": 1,
                      "type": "integer"
                    },
                    "total": {
                      "minimum": 0,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "data",
                    "page",
                    "limit",
                    "total"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "subjects"
        ]
      },
      "post": {
        "operationId": "SubjectsController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubjectDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "subjects"
        ]
      }
    },
    "/api/v1/subjects/{id}": {
      "get": {
        "operationId": "SubjectsController_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "^[1-9][0-9]{0,18}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "subjects"
        ]
      },
      "patch": {
        "operationId": "SubjectsController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "^[1-9][0-9]{0,18}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubjectDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "subjects"
        ]
      }
    },
    "/api/v1/subjects/{id}/archive": {
      "post": {
        "operationId": "SubjectsController_archive",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "^[1-9][0-9]{0,18}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "subjects"
        ]
      }
    },
...
  },
  "servers": [],
...
}